akakak

+++ pages/api/v1/db/deleteNews.ts
@@ -5,7 +5,15 @@ export default async function deleteNews(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+const { request } = req.headers;
   const { title, titleSlug, by, slug, sourceUrl, content } = req.body;
+
+  if(request == process.env.NEXT_SECRET_API_KEY) {
+    return res.status(401).json({
+      status: 401,
+      message: "Not authorized"
+    })
+  }
 
   const data = await deleteNewsHook({
     title: title,
